ANT-337 : GUI - New frozen item in Configure
It is planned to add a new item in the Configure menu.
Once clicked, this item will show the all the output files expected to be printed, as it is shown
in the inspector, when areas or links are selected.
The goal of this FUTURE WORK (not done in this commit) is to see at a glance the files that are printed, area by area and link by link.
This global view is not allowed with the inspector.
For now, we only add a frozen item in the Configure menu : "Filters on results by file"