ANT-344 - bug : no link output results visible in GUI for some study
For some study, when trying to view results in the Antares GUI, all results associated to links are missing.
More precisely, in the GUI's window related to results, the left panel supposed to show all clickable items for districts,
areas and links, the link items are missing.
When we look into the results directory, all link results are there, though.
This strange behavior is related to the fact that the study does not ask for synthesis results.
But these results (directory "mc-all") are required to list the links in the GUI's results window, so if synthesis
are not built, no link is detected.
Items related to areas are still displayed though, because file "<dir-results>/about-the-study/areas.txt" is used to this end.
We do the same with "<dir-results>/about-the-study/links.txt", which was not used up to now.