ANT-337 : conversion from any study format to 7.1 (filtering)
In a previous commit, file generaldata.ini's parameter "filtering" has been called "geographic-trimming".
What if we load a study whose format is 7.0 or less (with filtering key word) with an Antares 7.1 ?
Answer : as is, Antares 7.1 displays an error, though it should smoothly convert filtering into geographic-trimming with the same associated value.
This is the matter of the current commit.