ANT-360 : skipping to write results depending on selected variables
Since few commits, user can select the output variables she wants to be printed on disk during
an Antares study execution.
These results can roughly be divided in two parts : zonal results and link results.
If link variables are all deselected, then results directories associated to link results
turn out to be empty, and must not be created.
Same thing for zonal results if zonal variables are deselected.
This commit skip the creation of results directories that are empty due to the previous kind of deselection.