ANT-338 : when no geographic trimming at a study loading
In that case of geographic trimming turned to "None", we do not want the menu item "Configure > Geographic trimming..." to be frozen and disabled.
We want this item's sub-items ("Areas" or "Links") to be reachable and clickable, and when they are clicked,
the usual dialog box opens and displays the message "All geographic results will be printed on disk." in bold and in the right size.