ANT-338 : on the way to add links trimming (refactoring 1)
This is a refactoring step (step 1) :
Adding links trimming is very similar to adding areas trimming (already done at this stage).
We need to add a new dialog box and a new grid inside.
But we want to avoid duplicating code while doing this, so we need an inheritance tree that allow
factorizing code for both dialog boxes and grids.
This is the step 1 to this goal.