ANT-338 : on the way to add links trimming (refactoring 3)
Adding links trimming is very similar to adding areas trimming (already done at this stage).
We need to add a new dialog box and a new grid inside.
But we want to avoid duplicating code while doing this, so we need an inheritance tree that allow
factorizing code for both dialog boxes and grids.

This is the step 3 to this goal : Considering the areas trimming's grid as a child of a base class (geographic trimming grid)
that contains the common "trimming grid" code.