GUI : Refactoring of the scenario builder notebook's tabs creation
In order to introduce a new tab for hydro initial levels in the scenario builder notebook, we first refactor its tabs creation.
It will make architecture cleaner and ease the further addition.
A template function containing if... else... statement is replaced with a hierarchy of small polymorphic classes.
We get closer to the software architecture open/close principle.