scenario builder data : refactoring - splitting a class into more specialized classes.
class TSNumberRules (force time series to be chosen via its method apply()) contains treatments for every energy + demand.
A proof of that is the "if, else" or "switch" statements that can be found in this class's methods.
Here we split this class into as many classes as energy (+ demand).
This commit introduces a regression with the v7.1.
This will have to be fixed.