Scenario builder : data file values conversion
Lines from file "scenariobuilder.dat" are first turned into properties : basically key/value pairs, where the keys and the values are strings.
This pairs must be put into a data matrix : each key contains the information of the indices (x, y) where to put the value to in the matrix.
We should not convert the values before we know into what type they should be converted into.
However, before this commit, all values were integers, so values were converted naturally into integers, and too early during the execution (before execution point
is inside data classes).
But we plan to make this values doubles as well (for hydro levels).
So the purpose of this commit is to convert the values inside the data classes, that is when we know into what type should be converted the values.