Constraint summary : red bullet when weights are nil or clusters are disabled / must-run
In the panel "Binding constraint > Summary", when a constraint involves links or clusters with nil weights, the bullet related to this constraint should be red.
In the same way, if the constraint involves only nil weight links but non nil weight cluster, and if these clusters are disabled or must-run, the bullet related
to the constraint should be red.