Disabling sc builder menu item : actual item disabling
when "GUI > Simulation > Building mode" is turned to Automatic or Derated, the scenario builder is not used.
So in these cases, accessing to "GUI > Configure > MC scenario builder" has no sense and thus we disable this menu item and the access to the scenario builder.
Of course, in case "GUI > Simulation > Building mode" is Custom, the scenario builder can still be accessed.