[skip] Disabled clusters in scenario builder.
It make no sense that disabled clusters are explicitly chosen a scenario in THE ACTIVE RULE of the scenario builder.
So, for such thermal cluster, a warning is triggered only if it is related to a scenario in the active rule, not in any rule, like it was the case before this commit.

More over, such warnings should be fewer : instead of triggering a warning for each year a disabled cluster is needed (in the active rule), we send only one warning
listing the 10 first years (at most) the clusters is needed.