Scenario builder + hydro levels : bug fix
It is NOT correct to update hydro levels with scenario builder hydro levels at the beginning of each week.
Hydro levels are already updated at the beginning of each year in the hydro heuristic.
Priority of hydro levels from scenario builder over hot start is not affected by this small change.