Scenario builder + hydro levels : a further fix
Making the cost of wasting some hydro resources higher than the cost when levels are underneath the hydro low limit curve.

what was the problem ? Before this commit, start level being under the low limit is expensive.
So heuristic prevents from using the inflows (in order to produce) the first month and delays production to next month : the level increases.
It can be the same situation next month, and so on, until level reaches the low limit curves from underneath.
Once level is above the low limit curve, the same cost prevents from producing in order to reach down the initial level, meaning that we cannot
use all the inflows of the year : they is a waste of hydro resources.
We now consider that the waste of water use is more serious than not being under the level low limit.