Hydro heuristic policy : refactoring
The DAILY hydro heuristic policy (2nd step of the heuristic) involves an linear optimization.
The refactoring here involves separating the recovery + application of the optimization problem costs from its instantiation.
Doing that, we realized that .c files (containing necessarily C code so that compilation succeeds) are a real constraint when introducing a bit of C++ in their headers.
So they are now renamed as .cpp files.