Refactoring link upper banner : first try ==> failure at this stage
We tried to make the transmission capacities usage button and its behavior a unique object shared by
the 2 pages "Parameters" and "Transmission capacities".
This try compiles but fails at execution : when changing the value of the capacities usage on one page,
we cannot refresh this change on the other page. So we get back to where we used to be.