sc builder warnings too long
Warnings from scenario builder can be very long :
this happens when a cluster (thermal or renewable) is disabled and when some TS numbers are assigned to this cluster in the scenario builder.
As the cluster is disabled, Antares warns the user of all years where a TS number is set.
When this year collection is grows, the warning line grows accordingly.