study info file for statistics : the file's CSV writer now inherit form a base abstract class
StudyInfoWriter is now a base abstract class and its child StudyInfoCSVwriter is the writer to the CSV file.
Note that StudyInfoWriter is not exactly an interface, as it has data members supposed to be useful in any of its child writer,
especially a file content and a path to a file.