mps writer refactoring : defining an mps writer interface and consequences
Each way of printing MPS files (full historic, or-tools way, split format or none) is now isolated
inside a different class. Each of these ways can be removed or changed without impacting the client code.
A factory function contains the logic that leads to the creation of a particular MPS writer.