Simplify TS numbers draws : move permission to have 1 TS intra-modal where TS numbers are generated
2 commits ago, we removed ts numbers handling from class TimeSeries.
But it has side effect regarding results in case the energy is intra-modal : in this mode, having 1 TS only in an area is allowed.
The consequences of this allowance was handled by class TimeSeries. It had to be handled somewhere.
We move this code where TS numbers are generated, more specifically where intra-modal is handled.