Fail if OR-Tools solver is not found (#1851)
This change follows a discussion with @sylvlecl. According to him,
falling back to Sirius without OR-Tools is not a desirable behavior
since it can be misleading. For example, the user thinks they're using
XPRESS, but since it's not available (missing licence, misplaced env
variables, etc.), Antares Solver silently uses Sirius with a warning in
the logs instead.

The choice has been made to refuse running the simulation when the user
requests an unavailable solver, and to throw a nice error message

> [2023-12-27 10:17:30][solver][fatal] Can't use solver 'spx' : it does
not match any available OR-Tools solver. Possible choices are coin,
sirius, xpress.