Simplify TS numbers drawings, fix bug related to refresh & local thermal generation (#1752)
We intend to simplify and clarify the TS number management (specially
source file **timeseries-numbers.cpp**).
For this we did several things : 
- [x] For generated TS, separate resizing from generation, and call the
resizing before drawing TS number, so that we have the definitive TS
sizes before we draw the TS numbers : this is the **main leverage** for
TS number management simplifications and clarification.
- [x] At construction, the 1-d matrix **timeseriesNumbers** is given
zeros. This allows many simplifications and code reductions.
- [ ] For productions (only for **Load**, **Wind** and **Solar**) that
generate null TS (due to the configuration of their TS generator [see
prepro capacity]), raise a warning somewhere for user to know.