[CI] Don't use runner.name as a key for ubuntu's ccache (#1878)
`${{runner.name}}` refers to

> The name of the runner executing the job. This name may not be unique
in a workflow run as runners at the repository and organization levels
could use the same name.


[source](https://docs.github.com/en/actions/learn-github-actions/contexts#runner-context)

It may be too unique since after this change, ccache never finds cached
data (hit rate ~ 0%). This defeats the purpose of using a cache.

Using a fixed value may lead to a higher hit rate.