[CI] Add sirius lib to PATH  (#1725)
**Description**

On windows, running antares requires to have sirius DLL in the search
path for libraries.
In particular, unit tests linked with sirius need that.

The current implemented solution is a somewhat dirty workaround:
for each executable linked (possibly transitively) with sirius, we have
a special CMake macro which needs to be used to copy sirius DLL to the
same directory as the executable.

This requires developers to modify the build just to solve an
installation issue, and this for **each** executable that are linked
with sirius.

**Solution**
Instead, the PR proposes to add sirius installation directory to the
PATH variable.
Executing unit tests will now rely on normal windows DLL search
mechanism, and the developer only has to use the standard
`target_link_libraries` in CMake files.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>
Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>