Create & forward declare BasisStatus struct, reduce build time (#2044)
This is a wrapper around 2 vectors of OR-Tools-defined enums
BasisStatus. By forward-declaring this class in most places,
linear_solver.h is not included, which could avoid compiling it 29
times.

```
373174 ms: /home/leclercsyl/sources/antares/antares-simulator/src/solver/optimisation/include/antares/solver/optimisation/opt_structure_probleme_a_resoudre.h (included 124 times, avg 3009 ms), included via:
[...]
347162 ms: /home/leclercsyl/sources/antares/or-tools/install/debug/include/ortools/linear_solver/linear_solver.h (included 133 times, avg 2610 ms), included via:
  29x: opt_structure_probleme_a_resoudre.h 
```

A side effect is a better separation of responsibilities by having a
dedicated class. ~~I couldn't use `std::unique_ptr` with incomplete
type, but this should be fixable if need be.~~
[done](https://github.com/AntaresSimulatorTeam/Antares_Simulator/pull/2044/commits/c9fbd2502b6dfc0d79a0b078f5a0b0e1fe2e4baf)

Also, add unit 1 test for basis extraction.