Add logs to checks on number of columns, refactor (#2080)
Add more context to allow the user to correct invalid data.

Extract of logs after this PR
```
[2024-05-03 13:43:44][solver][error] Inconsistent number of columns for link fr / gb (direct-capacity: 3, indirect-capacity: 2)
[2024-05-03 13:43:44][solver][error] Please check that all series have the same number of columns or 1 column, or a combination of 1 column and the same number of columns
```

Also, small refactor to split the implementation of function
`TimeSeriesNumbers::CheckNumberOfColumns` in 2.

Using a `std::map` induces a slight change of behavior : unique keys. TS
with the same key will be erased. This is mostly good, for example if
multiple `AreaScratchpad` are created for a given area (parallel
simulations).