Parallel years cleaning : few removal of "numspace" (#2128)
See ticket [#1751](https://gopro-tickets.rte-france.com/browse/ANT-1751)

For now, just a first analyze of how we could do to isolate **years**
(from the simulation and what should not be inside a **year**).
We take advantage of this to make some removals of **numSpace**.

**Comments** : 
- **year job** and **simulation** are really intertwined, this is a
first and important reason why a **year** are not isolated.
For instance, when running a year of simulation, we call a function of
the **simulation** object, not of the **year** object.
- although what we first thought, **each year job creates a fresh new
weekly optimization solver object**. This object solves every week
problem, using the previous week's optimization solution as a base.
Before creating a solver object, a year first **destroys the previous
year's solver** (as well as the start base it holds).
- I've heard that CI tests for parallel (currently passed at night run)
don't generate the same output results as the same tests run in
sequential. We assumed that it was because of the bases. Indeed, if the
same simulation year is run starting from 2 different bases, it can lead
to different simulation results. But given the previous point, that's
not true : the simulation results should be identical whether we run a
simulation sequentially or not. We may have broken something in the
parallelism by not testing the it on CI.


**Goals** : 
We want to isolate year jobs to make cleaner parallelism.
Each year job should be designed so that it can be run in a thread
independently from any other year job.
In particular, a year job should hold its own memory spaces, and not
shared memory as done currently.

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>