Expose API [ANT-1158] (#1993)
Expose Antares Solver in a library. the library and target are called
AntaresAPI. This API expose one entry point to run a simulation on a
study. The return value of this is a data structure with the path to the
simulation and the weekly problems.

- New component API
- Create a new project "api_client" and corresponding tests to detect
API lib install regressions
- rename "utils" project in tests as "in-memory-study"
- promote this project at top level in tests. This allow in-memory-study
to be used for any tests
- create SimulationObserver. Object passed throughout all "frames" of
the simulation.

---------

Co-authored-by: Florian Omnès <florian.omnes@rte-france.com>
Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>