TS generator : refactor thermal part (#2202)
Aim of this PR : see the title.

What was done (see helping **gitHub** comments in this PR) : 
- simplifications : 
- [x] Standardizing TS generation with link TS : remove a constructor in
a class containing data for TS generation.
- [x] Remove use of **RuntimeInfos** in **main.cpp** (many code lines
used to be called there, just to get a random generator).
  - [x] Remove check on min stable power, useless.
- [x] Function that actually generates the TS no longer takes the TS to
be generated as argument, but returns it.
- [x] **TS generator** code no longer contains code related to whether
we should skip the TS writing on disk. It writes, that'all. This kind of
code belongs to **solver**, not to **TS generator**.
  - [x] Use **std::filesystem** whenever possible instead of **Yuni**
  
- Unwanted behavior correction : 
- [x] Execute links or thermal TS generation code only depending on user
requirements : indeed, thermal TS generation still needs the whole study
to be loaded, but link TS generation does not (it retrieves from study
only the required data). Being forced to load the study if only link TS
generation is required is a clear performance issue. We avoid that here.
- [x] When data files required to generate TS (modulation files & prepro
files) don't exist, we used to ignore it (this can lead to a crash). We
now raise an error before TS generation time comes.
- Moving code : _to be completed_

---------

Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>