Fix/remove useless fwd decl (#13)
* Collect hydro validation errors (#2204)

collect *n* hydro validation errors (10 per area ) before exiting

---------

Co-authored-by: Florian Omnès <florian.omnes@rte-france.com>

* Remove actions dependencies using node js 16 (#2215)

> Node.js 16 actions are deprecated. Please update the following actions
to use Node.js 20: notiz-dev/github-action-json-property@release. For
more information see:
https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.

---------

Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>

* Move TS number print (#2228)

Purpose : moving the code that prints the TS numbers on disk : due to
recent changes it was moved after the loop through MC years. So, we have
to wait until simulation ends to know which TS numbers where drawn. So
we move that code back.

* Infeability analyzer : renaming (#2225)

Purpose : as the title says

This PR is attached to ticket [ticket
ANT-1825](https://gopro-tickets.rte-france.com/browse/ANT-1825).
Some improvements were made or tried by taking care of the ticket.
The result is this PR.

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>

* Always run clang-format on PR (#2230)

* Add changelog for minor versions (#2229)

- v8.8.6
- v8.6.8

---------

Co-authored-by: Abdoulbari Zaher <32519851+a-zakir@users.noreply.github.com>
Co-authored-by: Jason Maréchal <45510813+JasonMarechal25@users.noreply.github.com>

* Infeasibility analyzer : small simplifications (#2226)

Purpose : as the title says

This PR is attached to ticket [ticket
ANT-1825](https://gopro-tickets.rte-france.com/browse/ANT-1825).
Some improvements were made or tried by taking care of the ticket.
The result is this PR.

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>

* Fix compile, add boost header in CMakelists (#2233)

* STS: Withdrawal efficiency [ANT-1862] (#2223)

Add a efficiencywithdrawal property to short term storage
This mimics the already existing efficiency for injection

* Infeasibility analyzer : HydroPower constraint (#2227) [ANT-1825]

Add constraint that can possibily lead to infeasibilities. HydroPower is the constraint on the sum of generated hydro energy over the week for a given area. The provided RHS in some cases makes the problem infeasible.

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>

* Compilation warnings (#2237)

close  #2236

* Version 9.2 (#2240)

* Fix sonarcloud job (#2246)

* Adequacy Patch regression [ANT-1845] (#2235)

* 9.2 rc 1 (#2247)

* Keep using node js 16 on centos CI (#2248)

* Infeasibility more cleaning (#2231)

We're on the road to make **infeasibility analyzer** more changeable,
more precisely when adding a constraint type to the list of constraints
to be detected in case of infeasibity.
This PR is a step towards this purpose.  
It contains some heterogeneous simplifications and renaming.
In order to ease the review, some comments were added in this PR.

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>

* Document clang-format (#2243)

Add version number to avoid confusion.

* Link TS generation : splitting into multiple files (#2171)

This PR aims at moving some code about link TS generation, from **main**
program to new source files.
It follows PR #2155 

To be done : 

- [x] move headers (*.h) to the right place
- [x] update this branch with its base branch (@flomnes made a change on
base branch)

We may also take advantage of this PR to : 
- [ ] Thermal TS generation :
- [ ] avoid loading the **whole study** when **no** thermal TS
generation is required (this is currently the case)
- [ ] move implementation details about thermal TS generation from
**main** program to new source files
- [ ] Separate loading / extracting data from study and generate the TS

---------

Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>
Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>

* Explain why MPS are named when the problem is infeasible (#2250)

* Add tests in CI for version 9.2 (#2241)

Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>

* Rename test-platform to os for windows CI (#2253)

* Fix bug hydro heuristic with mingen (ANT-1825) (#2258)

Removing lines that create infeasibility and that are not necessary.

---------

Co-authored-by: Juliette-Gerbaux <juliette.gerbaux_externe@rte-france.com>
Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>

* Remove useless fwd declaration

---------

Co-authored-by: Abdoulbari Zaher <32519851+a-zakir@users.noreply.github.com>
Co-authored-by: Florian Omnès <florian.omnes@rte-france.com>
Co-authored-by: payetvin <113102157+payetvin@users.noreply.github.com>
Co-authored-by: guilpier-code <62292552+guilpier-code@users.noreply.github.com>
Co-authored-by: Florian OMNES <26088210+flomnes@users.noreply.github.com>
Co-authored-by: Jason Maréchal <45510813+JasonMarechal25@users.noreply.github.com>
Co-authored-by: Juliette-Gerbaux <130555142+Juliette-Gerbaux@users.noreply.github.com>
Co-authored-by: Juliette-Gerbaux <juliette.gerbaux_externe@rte-france.com>