API modeler - 1.1.b (#2391)
The goal of this PR is to introduce and unit test class
**LinearProblemBuilder**.
This class is an orchestration class, and therefore makes other classes
collaborate.
What was done : 
- Remove **LinearProblemBuilder::solve()** : 
this method was planned in the specifications, but it extends the
builder responsibility scope.
Once modified by the builder, the problem can be solved by calling its
own **solve()** method.
- Add methods **numVariables()** and **numConstraints()** to base class
**ILinearProblem** for unit tests reason.
- **CMakeLists.txt** file for modeler API was split into 2 **\*.cmake**
files
- **Trial for a dependency simplification** :
In the original / specified architecture, **LinearProblemBuilder**
collaborates with classes :
  - **LinearProblemData**
  - **ILinearProblem** (interface for LP)
- **LinearProblemFiller** : this class also depends on previous classes

Here, we allowed ourselves to reduce dependencies of this builder class
(is this simplification relevant ?) :
class **LinearProblemBuilder** currently only depends on
**LinearProblemFiller**, while fillers keeps its planned dependencies on
**LinearProblemData** and **ILinearProblem**.
Indeed, any filler is in charge to modify problem (using LP data), the
but the builder only manipulates fillers.
- **Mocking fillers** : fillers are the entities responsible for
creating variables and constraints in the linear problem. To do that,
they need the LP itself as well as the LP data (see fillers
constructor). We created an abstract base filler. All concrete filler
inherit from it. Class **LinearProblemBuilder** is constructed from a
vector of fillers.
- Several unit tests were created for class **LinearProblemBuilder**

---------

Co-authored-by: Abdoulbari Zaher <32519851+a-zakir@users.noreply.github.com>