Remix multi storage : passing pmax to algorithm by copy (see commit message)
In many cases, pmax does not have a steady address during preparation of algorithm input.
Indeed, if passed by reference (const or not), it can be destroyed / corrupted before being used by algorithm.

More precisely, it's passed to class StorageForRemixWithLevels's constructor and stored in the associated instance.
When the list of StorageForRemixWithLevels is built, we leave the scope where pmax was constructed : nothing prevents
it's destroyed / corrupted.